Added casting so that the radius channel works with GPU tensors.

+++ CoordConv.py
@@ -101,7 +101,7 @@ class AddCoords(nn.Module):
             yy_channel.type_as(input_tensor)], dim=1)
 
         if self.with_r:
-            rr = torch.sqrt(torch.pow(xx_channel - 0.5, 2) + torch.pow(yy_channel - 0.5, 2))
+            rr = torch.sqrt(torch.pow(xx_channel.type_as(input_tensor) - 0.5, 2) + torch.pow(yy_channel.type_as(input_tensor) - 0.5, 2))
             ret = torch.cat([ret, rr], dim=1)
 
         return ret
